% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{all_permutations}
\alias{all_permutations}
\title{All permutations used to compute the exact distributions of circular tests}
\usage{
all_permutations(n = 2, m = 2)
}
\arguments{
\item{\code{n}}{total of frequencies}

\item{\code{m}}{number of spacings}
}
\value{
A \code{matrix} with all possible permutations
}
\description{
All permutations used to compute the exact distributions of circular tests
}
\examples{
all_permutations(2,2)
all_permutations(4,3)
}
