% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{circular_test}
\alias{circular_test}
\title{Two sample Test of Homogeneity}
\usage{
circular_test(x, y, test = "dixon", alpha = 0.05, B = NULL,
  type = "exact")
}
\arguments{
\item{\code{x}}{first sample}

\item{\code{y}}{second sample}

\item{\code{test}}{considered test (\code{dixon} for Dixon test (default);
\code{ww} for Wheeler-Watson; \code{wilcox} for Wilcoxon test; \code{rao}
for Rao test; \code{vdw} for van der Waerden test; \code{savage}
for Savage test}

\item{\code{alpha}}{significance level (default = 0.05)}

\item{\code{B}}{number of bootstrap replications}

\item{\code{type}}{method to compute pvalues (available methods: \code{exact} for exact computation (default) which is appropriate for small sample sizes;
\code{mc} for approximation based on Monte-carlo simulations)}
}
\value{
A list with the following structure:
\describe{
 \item{cv}{a \code{list} containing the results of the exact distribution (\code{NULL} if \code{type = "mc"}), see function \code{get_critical_values} for details}
 \item{mc}{a \code{list} containing the results of the approximated distribution obtained by simulation (\code{NULL} if \code{type = "exact"}), see function \code{MC_pvalue} for details}
 \item{B}{number of bootstrap replications}
 \item{alpha}{significance level}
 \item{test}{the considered test}
 \item{stat}{observed test statistic}
 \item{spacings}{a \code{list} containing the observed spacings, see function \code{compute_Sk} for details}
 \item{x}{expression deparsing of the first dataset}
 \item{y}{expression deparsing of the second dataset}
 }
}
\description{
This function allows to perform various nonparametric test for homogeneity on two samples of circular data.
}
\examples{
# Load dataset
data(pigeons)
# Dixon test (exact pvalue)
circular_test(pigeons$experimental, pigeons$control)

# Dixon test (approximated pvalue)
circular_test( pigeons$experimental, pigeons$control, type = "mc")
}
