% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{compute_Sk}
\alias{compute_Sk}
\title{Compute spacing frequencies}
\usage{
compute_Sk(x, y)
}
\arguments{
\item{\code{x}}{first sample}

\item{\code{y}}{second sample}
}
\value{
A list with the following structure:
\describe{
 \item{Sk}{a \code{vector} containing the spacing frequencies}
 \item{n}{total of frequencies}
 \item{m}{number of spacings}
}
}
\description{
Compute spacing frequencies
}
\examples{
data(pigeons)
compute_Sk(pigeons$control, pigeons$experimental)
}
