% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{dixon}
\alias{dixon}
\title{Dixon test statistic based on spacing frequencies}
\usage{
dixon(Sk)
}
\arguments{
\item{\code{Sk}}{spacing frequencies}
}
\value{
Dixon test statistic
}
\description{
Dixon test statistic based on spacing frequencies
}
\examples{
data(pigeons)
spacings = compute_Sk(pigeons$control, pigeons$experimental)
dixon(spacings$Sk)
dixon(c(1,0,0,0,1,2))
}
