% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{get_bracket}
\alias{get_bracket}
\title{Compute Bracketing Values}
\usage{
get_bracket(x, alpha)
}
\arguments{
\item{x}{A \code{vector} of test statistics}

\item{alpha}{Significance level}
}
\value{
Bracketing values (c1, c2) corresponding to significance levels (p1, p2)
}
\description{
Compute bracketing values (c1, c2) corresponding to significance levels (p1, p2)
based on the significance level alpha (such that p1 > alpha > p2)
}
\keyword{internal}
