% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{MC_pvalue}
\alias{MC_pvalue}
\title{Monte-carlo simulation function to compute pvalues}
\usage{
MC_pvalue(stat, n, m, test, B = 10^4, seed = 1982)
}
\arguments{
\item{\code{stat}}{observed test statistics}

\item{\code{n}}{total of frequencies}

\item{\code{m}}{number of spacings}
}
\value{
A list with the following structure:
\describe{
 \item{dist}{a \code{vector} containing the approximate distribution of the test statistic}
 \item{n}{total of frequencies}
 \item{m}{number of spacings}
 \item{test}{the considered test}
 }
}
\description{
This function approximate by Monte-carlo simulation the distribution
of a given test statistic under the H0.
}
