% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pigeons}
\alias{pigeons}
\title{Homing pigeons dataset}
\format{A list:
\describe{
  \item{control}{release flight direction of the pigeons (control group)}
  \item{experimental}{release flight direction of the pigeons (experimental group)}
}}
\source{
Schmidt-Koenig, K. (1958).
}
\usage{
pigeons
}
\description{
A dataset containing the release flight direction of the pigeons
from the two groups were collected in different experiments conditions.
}
\keyword{datasets}
