% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{plot.circular_test}
\alias{plot.circular_test}
\title{Plot null distribution associated to two sample test of Homogeneity}
\usage{
\method{plot}{circular_test}(x, cex.main = 1)
}
\arguments{
\item{x}{A \code{circular_test} object}
}
\value{
Plots the (exact or approximated) distribution of the test statistics under H0.
}
\description{
Plots the (exact or approximated) distribution of the test statistics under H0.
}
\examples{
data(pigeons)
plot(circular_test(pigeons$experimental, pigeons$control))
}
\keyword{internal}
