% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{plot.critical_values}
\alias{plot.critical_values}
\title{Plot null distribution associated to two sample test of Homogeneity}
\usage{
\method{plot}{critical_values}(x, main = NULL, cex.main = 1)
}
\arguments{
\item{x}{A \code{critical_values} object}
}
\value{
Plots the (exact or approximated) distribution of the test statistics under H0.
}
\description{
Plots the (exact or approximated) distribution of the test statistics under H0.
}
\examples{
plot(get_critical_values(6,8))
}
\keyword{internal}
