% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{print.circular_test}
\alias{print.circular_test}
\title{Print Two sample Test of Homogeneity}
\usage{
\method{print}{circular_test}(out)
}
\arguments{
\item{x}{A \code{circular_test} object}
}
\value{
Prints the results of a two sample test of homogeneity
}
\description{
Prints the results of a two sample test of homogeneity
}
\examples{
# Load dataset
data(pigeons)
# Dixon test (exact pvalue)
circular_test(pigeons$experimental, pigeons$control)
}
\keyword{internal}
