% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crit_val.R
\name{print.critical_values}
\alias{print.critical_values}
\title{Print Bracketing Values for Two Sample Test of Homogeneity}
\usage{
\method{print}{critical_values}(x)
}
\arguments{
\item{x}{A \code{critical_values} object}
}
\value{
Prints bracketing values and corresponding to significance levels
for a two sample test of homogeneity
}
\description{
Prints bracketing values and corresponding to significance levels
for a two sample test of homogeneity
}
\examples{
# Load dataset
get_critical_values(6,8)
}
\keyword{internal}
