% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{rao}
\alias{rao}
\title{Rao test statistic based on spacing frequencies}
\usage{
rao(Sk)
}
\arguments{
\item{\code{Sk}}{spacing frequencies}
}
\value{
Rao test statistic
}
\description{
Rao test statistic based on spacing frequencies
}
\examples{
data(pigeons)
spacings = compute_Sk(pigeons$control, pigeons$experimental)
rao(spacings$Sk)
rao(c(1,0,0,0,1,2))
}
