% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{savage}
\alias{savage}
\title{Savage test statistic based on spacing frequencies}
\usage{
savage(Sk)
}
\arguments{
\item{\code{Sk}}{spacing frequencies}
}
\value{
Savage test statistic
}
\description{
Savage test statistic based on spacing frequencies
}
\examples{
data(pigeons)
spacings = compute_Sk(pigeons$control, pigeons$experimental)
savage(spacings$Sk)
savage(c(1,0,0,0,1,2))
}
