% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{van.der.Waerden}
\alias{van.der.Waerden}
\title{van der Waerden test statistic based on spacing frequencies}
\usage{
van.der.Waerden(Sk)
}
\arguments{
\item{\code{Sk}}{spacing frequencies}
}
\value{
van der Waerden test statistic
}
\description{
van der Waerden test statistic based on spacing frequencies
}
\examples{
data(pigeons)
spacings = compute_Sk(pigeons$control, pigeons$experimental)
van.der.Waerden(spacings$Sk)
van.der.Waerden(c(1,0,0,0,1,2))
}
