% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{wheeler_watson}
\alias{wheeler_watson}
\title{Wheeler-Watson test statistic based on spacing frequencies}
\usage{
wheeler_watson(Sk)
}
\arguments{
\item{\code{Sk}}{spacing frequencies}
}
\value{
Wheeler-Watson test statistic
}
\description{
Wheeler-Watson test statistic based on spacing frequencies
}
\examples{
data(pigeons)
spacings = compute_Sk(pigeons$control, pigeons$experimental)
wheeler_watson(spacings$Sk)
wheeler_watson(c(1,0,0,0,1,2))
}
