% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{wilcoxon}
\alias{wilcoxon}
\title{Wilcoxon test statistic based on spacing frequencies}
\usage{
wilcoxon(Sk)
}
\arguments{
\item{\code{Sk}}{spacing frequencies}
}
\value{
Wilcoxon test statistic
}
\description{
Wilcoxon test statistic based on spacing frequencies
}
\examples{
data(pigeons)
spacings = compute_Sk(pigeons$control, pigeons$experimental)
wilcoxon(spacings$Sk)
wilcoxon(c(1,0,0,0,1,2))
}
